pyproject.toml
[tool.poetry]
name = "notificationservice"
version = "0.1.0"
description = ""
authors = ["ZhuraTown <Swiglek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
pika = "^1.3.2"
websockets = "^14.2"
pydantic-settings = "^2.7.1"
click = "^8.1.8"
aio-pika = "^9.5.4"
alembic = "^1.14.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
uvicorn = "^0.34.0"
passlib = "^1.7.4"
python-jose = "^3.4.0"
bcrypt = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
